moved Param and Seg into seperate hpp files
